requirements.txt
aiogram==3.22
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.9
jinja2>=3.1
redis>=5.0
sqlalchemy>=2.0
alembic>=1.13
python-dotenv>=1.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
aiohttp>=3.9.0
requests>=2.31.0
